+++ lib/lightning_web/live/components/icon.ex
@@ -13,6 +13,10 @@ defmodule LightningWeb.Components.Icon do
 
   def runs(assigns), do: Heroicons.rectangle_stack(assigns)
 
+  def pencil(assigns), do: Heroicons.pencil(assigns)
+
+  def exclamation_circle(assigns), do: Heroicons.exclamation_circle(assigns)
+
   def settings(assigns), do: Heroicons.cog_8_tooth(assigns)
 
   def dataclips(assigns), do: Heroicons.cube(assigns)